Embed sections into course
